Added Hyper-V Utilities "Backup" and "Shutdown"
Added powershell scripts to handle backing up and shutting down a Hyper-V system.
HyperVBackup is intended to be run on a storage host against a remote Hyper-V server and utilizes the Hyper-V powershell module.
HyperVSafeShutdown is intended for use on a Hyper-V host directly to avoid such issues as "Stopping Hyper-V Virtual Machine Management Service" when the host is Rebooted or Shut Down.